+++ api/CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to the **Prowler API** are documented in this file.
 - Support Prowler ThreatScore for the K8S provider [(#9235)](https://github.com/prowler-cloud/prowler/pull/9235)
 - Enhanced compliance overview endpoint with provider filtering and latest scan aggregation [(#9244)](https://github.com/prowler-cloud/prowler/pull/9244)
 - New endpoint `GET /api/v1/overview/regions` to retrieve aggregated findings data by region [(#9273)](https://github.com/prowler-cloud/prowler/pull/9273)
+- New endpoint `GET /api/v1/overview/findings_severity_timeseries` to retrieve aggregated timeseries severities in granular way and Add new Index to `ScanSummary` model [(9307)](https://github.com/prowler-cloud/prowler/compare/PROWLER-25-finding-severity-over-time-component-api)
 
 ### Changed
 - Optimized database write queries for scan related tasks [(#9190)](https://github.com/prowler-cloud/prowler/pull/9190)